# swagger/spec.json
{"openapi":"3.0.0","info":{"title":"Smart city API","version":"1.0.0"},"paths":{"/vehicule/{immatriculation}":{"get":{"security":[{"bearerAuth":[]}],"summary":"récupere un vehicule","tags":["Vehicule"],"parameters":[{"in":"path","description":"id du vehicule a récupérer","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehiculeId"}}}}],"responses":{"200":{"$ref":"#/components/responses/getVehicule"},"400":{"description":"mauvais format"},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"description":"acces non autorisé"},"404":{"description":"vehicule non trouvé"},"500":{"description":"erreur serveur"}}},"delete":{"security":[{"bearerAuth":[]}],"tags":["Vehicule"],"summary":"supprime un vehicule","parameters":[{"in":"path","description":"id du vehicule a supprimer","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehiculeId"}}}}],"responses":{"200":{"description":"Véhicule supprimer"},"400":{"description":"mauvais format"},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"description":"acces non autorisé"},"404":{"description":"vehicule non trouvé"},"500":{"description":"erreur serveur"}}}},"/vehicule":{"post":{"security":[{"bearerAuth":[]}],"summary":"crée un vehicule","tags":["Vehicule"],"requestBody":{"description":"véhicule à rajouter","required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehiculeToAdd"}}}},"responses":{"201":{"$ref":"#/components/responses/addVehicule"},"400":{"description":"mauvais format"},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"description":"acces non autorisé"},"500":{"description":"erreur serveur"}}},"patch":{"security":[{"bearerAuth":[]}],"summary":"modifie un vehcule","tags":["Vehicule"],"requestBody":{"description":"véhicule à modifier","required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehiculeToUpdate"}}}},"responses":{"204":{"$ref":"#/components/responses/updateVehicule"},"400":{"description":"mauvais format"},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"description":"acces non autorisé"},"404":{"description":"vehicule non trouvé"},"500":{"description":"erreur serveur"}}},"get":{"tags":["Vehicule"],"summary":"Recherche de véhicules","parameters":[{"in":"query","name":"search","schema":{"$ref":"#/components/schemas/VehiculeSearch"},"description":"Paramètres de recherche combinés"}],"responses":{"200":{"description":"Liste des véhicules correspondant","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Vehicule"}}}}},"400":{"description":"mauvais format"},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"description":"acces non autorisé"},"404":{"description":"vehicule non trouvé"},"500":{"description":"erreur serveur"}}}}},"components":{"schemas":{"Vehicule":{"type":"object","properties":{"id":{"type":"integer"},"immatriculation":{"type":"string"},"nbPlace":{"type":"integer"},"utilisateurId":{"type":"integer"}}},"Search_params":{"type":"object","properties":{"offset":{"type":"integer","minimum":0,"description":"Nombre de résultats à ignorer (pagination)","example":0},"limit":{"type":"integer","minimum":0,"description":"Nombre maximum de résultats à retourner","example":10},"fields":{"type":"string","description":"Liste des champs à retourner, séparés par des virgules","example":"id,name,email"},"withCount":{"type":"boolean","description":"Indique si le total des résultats doit être retourné","example":true}}},"VehiculeToAdd":{"type":"object","properties":{"immatriculation":{"type":"string","description":"Numéro d'immatriculation du véhicule","example":"1-ABC-123"},"nbPlaceMax":{"type":"integer","description":"Nombre maximum de places dans le véhicule","example":5},"utilisateurId":{"type":"integer","description":"Identifiant de l'utilisateur propriétaire","example":42}},"required":["immatriculation","nbPlaceMax","utilisateurId"]},"VehiculeToUpdate":{"type":"object","properties":{"vehiculeId":{"type":"integer","description":"Identifiant du véhicule à mettre à jour","example":1},"immatriculation":{"type":"string","description":"Numéro d'immatriculation","example":"1-ABC-123"},"nbPlaceMax":{"type":"integer","description":"Nombre maximum de places","example":5},"utilisateurId":{"type":"integer","description":"Identifiant de l'utilisateur propriétaire","example":42}},"required":["vehiculeId"]},"VehiculeId":{"type":"object","properties":{"id":{"type":"integer","description":"Identifiant du véhicule","example":1}},"required":["id"]},"VehiculeOptional":{"type":"object","properties":{"vehicule_id":{"type":"integer","description":"Identifiant du véhicule","example":1},"immatriculation":{"type":"string","description":"Immatriculation","example":"1-ABC-123"},"nbPlaceMax":{"type":"integer","description":"Nombre maximum de places","example":5},"utilisateurId":{"type":"integer","description":"Identifiant du propriétaire","example":42}},"additionalProperties":false},"VehiculeSearch":{"allOf":[{"$ref":"#/components/schemas/VehiculeOptional"},{"$ref":"#/components/schemas/Search_params"}]}},"responses":{"addVehicule":{"description":"add the given vehicule and return his identifiant","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vehicule"}}}},"getVehicule":{"description":"get the given vehicule","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehiculeId"}}}},"updateVehicule":{"description":"update the given vehicule","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehiculeToUpdate"}}}},"deleteVehicule":{"description":"delete the given vehicule","content":{"application/json":{"schema":{"type":"object","properties":{"vehiculeId":{"type":"integer"}}}}}},"searchVehicule":{"description":"Liste des véhicules correspondant aux paramètres de recherche","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Vehicule"}}}}},"UnauthorizedError":{"description":"JWT is missing or invalid"}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"tags":[]}